more refactoring, fixes to workflow, renaming
