+++ tests/test_neo4j/neo4j_connect.py
@@ -2,7 +2,7 @@
 # pylint: disable=broad-exception-caught broad-exception-raised
 import os
 
-from api.graphs.neo4j_graph import Neo4JGraph
+from app.graphs.neo4j_graph import Neo4JGraph
 from dotenv import load_dotenv
 
 load_dotenv()
