+++ tests/test_neo4j/test_mcq_generator_with_neo4j.py
@@ -3,10 +3,10 @@ import os
 from typing import Generator
 
 import pytest
-from api.graphs.mcq_graph import MCQGraph
-from api.graphs.neo4j_graph import Neo4JGraph
-from api.models import MCQ
-from api.utils.mcq_generator import MCQGenerator
+from app.core.mcq_generator import MCQGenerator
+from app.graphs.mcq_graph import MCQGraph
+from app.graphs.neo4j_graph import Neo4JGraph
+from app.models import MCQ
 from tests.test_neo4j.neo4j_connect import URI
